Upload Imported Data Analysis Function
This function handles raw data from the device to produce the following indicators attached to which night they are measured: basal body temperature, average heart rate, and most recent heart rate. The output is an updated indicator python pandas dataframe for use in fertile window and cycle predication as well as a .csv file for use in the application display.